# openapi/ancf.v1.yaml
openapi: "3.0.3"
info:
  title: ANCF Commerce API v1
  description: |
    Agent-Native Commerce Framework (ANCF) v1 API.

    ANCF is a Zero-Frontend Commerce protocol where Agents discover shops,
    search products, obtain backend-authoritative quotes, and submit signed
    checkout intents. The backend is authoritative for all trusted state:
    pricing, inventory, ledger, minting, and provisioning.

    ## Agent Requirements

    Agents must:
    - Fetch and validate `/.well-known/agent-rules.json` before any commerce action.
    - Verify manifest signature, schema, and firmware SRI.
    - Call `/quote` after `/search`; never use search prices for checkout.
    - Send `Idempotency-Key` for all mutating APIs.
    - Require human confirmation before real asset movement.
    - Never call mint/redemption APIs unless the user explicitly chooses them.

    ## Security

    - All mutating APIs require an `Idempotency-Key` header.
    - `checkout/commit` requires a valid wallet signature over the canonical signable payload.
    - `X-API-Key` header is required for API authentication.
    - The backend verifies: quote freshness, idempotency, wallet signature,
      inventory availability, vUSDC balance, and Agent session policy.
  version: "1.0.0"
  contact:
    name: ANCF Team
    url: https://github.com/ancf-commerce/ancf
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.yourshop.com
...
paths:
...
  /api/v1/wallet/deposit-intents:
    post:
      tags: [Wallet]
      summary: Create a deposit intent for vUSDC minting
      description: |
        Request a reserve address to deposit USDC (or another accepted asset).
        The returned `deposit_intent_id` and `reserve_address` should be used
        to complete the on-chain transfer. Once the chain adapter confirms the
        deposit, the mint-service credits the user's vUSDC balance.

        Agents must never call this API unless the user explicitly chooses deposit.
      operationId: createDepositIntent
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: Unique idempotency key
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositIntentRequest'
            example:
              wallet: "USER_WALLET"
              network: "solana-mainnet"
              asset_symbol: "USDC"
              amount_minor: "100000000"  # 100 USDC at 6 decimals
      responses:
        "200":
          description: Deposit intent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositIntentResponse'
              example:
                deposit_intent_id: "deposit_01J..."
                reserve_address: "RESERVE_ADDRESS_BASE58"
                memo: "deposit_01J..."
                expected_asset: "USDC"
                expected_amount_minor: "100000000"
                expires_at: "2026-06-04T01:00:00Z"
                network: "solana-mainnet"
                status: "pending"
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authenticating requests. Required for all endpoints except /health.
  schemas:
...
    DepositIntentRequest:
      type: object
      required: [wallet, network, asset_symbol, amount_minor]
      properties:
        wallet:
          type: string
          minLength: 1
        network:
          type: string
          enum: [solana-mainnet, sonic-l2, solana-devnet]
        asset_symbol:
          type: string
          minLength: 1
        amount_minor:
          type: string
          pattern: '^[0-9]+$'
        payment_rail:
          type: string
    DepositIntentResponse:
      type: object
      required: [deposit_intent_id, reserve_address]
      properties:
        deposit_intent_id:
          type: string
          pattern: '^deposit_'
        reserve_address:
          type: string
          minLength: 1
        memo:
          type: string
        expected_asset:
          type: string
        expected_amount_minor:
          type: string
          pattern: '^[0-9]+$'
        expires_at:
          type: string
          format: date-time
        network:
          type: string
          enum: [solana-mainnet, sonic-l2, solana-devnet]
        status:
          type: string
          enum: [pending, confirmed, credited, failed, expired]
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  reason:
                    type: string
            request_id:
              type: string
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "INVALID_PARAMETER"
              message: "Validation failed for field 'wallet'"
              request_id: "req_01J..."
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Missing or invalid X-API-Key header"
              request_id: "req_01J..."
security:
  - ApiKeyAuth: []